Landing Page
Our page should be a mix of "index.html" & "index_landing.html". The layouts and description given in the template have a good resemblance with our thought process when we will provide the template to the users. Although ours will be instantly editable without writing a single line of code.